+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Data Transfer Object classes generation from swagger spec
 - CacheableDispatcherFactoryProxy
+- JsonApiResponseFactory::createRelationshipResponse()
 
 ## [0.0.13] - 2021-01-27
 ### Added